Use fig if available.

+++ docker_rails_app.sh
@@ -16,33 +16,37 @@ function fix_file_permissions {
   find . \! -user dev -print0 | xargs -0 -I % sh -c 'sudo chmod g+w "%"; sudo chown dev:dev "%"'
 }
 
-if [ -n "$elasticsearch" ]; then
-  elasticsearch_directory='/home/vagrant/elasticsearch'
-  mkdir -p $elasticsearch_directory
-  if [ ! -f "$elasticsearch_directory/elasticsearch.yml" ]; then
-    echo "path:" > $elasticsearch_directory/elasticsearch.yml
-    echo "  logs: /data/log" >> $elasticsearch_directory/elasticsearch.yml
-    echo "  data: /data/data" >> $elasticsearch_directory/elasticsearch.yml
+if [ ! -e fig.yml ]; then
+
+  if [ -n "$elasticsearch" ]; then
+    elasticsearch_directory='/home/vagrant/elasticsearch'
+    mkdir -p $elasticsearch_directory
+    if [ ! -f "$elasticsearch_directory/elasticsearch.yml" ]; then
+      echo "path:" > $elasticsearch_directory/elasticsearch.yml
+      echo "  logs: /data/log" >> $elasticsearch_directory/elasticsearch.yml
+      echo "  data: /data/data" >> $elasticsearch_directory/elasticsearch.yml
+    fi
+    docker_do start elasticsearch || docker_do run -d --name=elasticsearch -p 9200:9200 -p 9300:9300 -v $elasticsearch_directory:/data dockerfile/elasticsearch /elasticsearch/bin/elasticsearch -Des.config=/data/elasticsearch.yml
+    extra="$extra --link elasticsearch:es -e ELASTICSEARCH_URL=es:9200"
   fi
-  docker_do start elasticsearch || docker_do run -d --name=elasticsearch -p 9200:9200 -p 9300:9300 -v $elasticsearch_directory:/data dockerfile/elasticsearch /elasticsearch/bin/elasticsearch -Des.config=/data/elasticsearch.yml
-  extra="$extra --link elasticsearch:es -e ELASTICSEARCH_URL=es:9200"
-fi
 
-# docker inspect --format=' ' $db
-# if [ $? -ne 0 ]; then
-if grep -q mysql $directory/config/database.yml; then
-  db=mysql
-  db_username='admin'
-  docker_do start $db || docker_do run -d --name=mysql -p 3306:3306 -e MYSQL_PASS="$db_password" tutum/mysql
-fi
-if grep -q postgres $directory/config/database.yml; then
-  db=postgresql
-  docker_do start $db || docker_do run -d --name=postgresql -p 5432:5432 -e POSTGRESQL_USER=$db_username -e POSTGRESQL_PASS=$db_password kamui/postgresql
-fi
+  # docker inspect --format=' ' $db
+  # if [ $? -ne 0 ]; then
+  if grep -q mysql $directory/config/database.yml; then
+    db=mysql
+    db_username='admin'
+    docker_do start $db || docker_do run -d --name=mysql -p 3306:3306 -e MYSQL_PASS="$db_password" tutum/mysql
+  fi
+  if grep -q postgres $directory/config/database.yml; then
+    db=postgresql
+    docker_do start $db || docker_do run -d --name=postgresql -p 5432:5432 -e POSTGRESQL_USER=$db_username -e POSTGRESQL_PASS=$db_password kamui/postgresql
+  fi
 
-if [ -z "$db_link" ]; then
-  db_link="--link $db:db -e DB_NAME=$app -e DB_USERNAME=$db_username -e DB_PASSWORD=$db_password"
-fi
+  if [ -z "$db_link" ]; then
+    db_link="--link $db:db -e DB_NAME=$app -e DB_USERNAME=$db_username -e DB_PASSWORD=$db_password"
+  fi
+
+fi # fig
 
 command="$1"
 shift
@@ -52,7 +56,11 @@ if [ $command = "b" ]; then # build
     fullpath=$(readlink -f Dockerfile)
     rm Dockerfile
     cp $fullpath Dockerfile
-    docker_do build --force-rm -t $app .
+    if [ -e fig.yml ]; then
+      fig build
+    else
+      docker_do build --force-rm -t $app .
+    fi
     rm Dockerfile
     ln -s $fullpath Dockerfile
   else
@@ -61,8 +69,12 @@ if [ $command = "b" ]; then # build
 fi
 
 if [ $command = "bundle" ]; then # bundle
-  docker_do run -i --rm $db_link $code_volume $extra $app /usr/local/bin/bundle "$@"
-  fix_file_permissions
+  if [ -e fig.yml ]; then
+    fig run --rm web bundle $@
+  else
+    docker_do run -i --rm $db_link $code_volume $extra $app /usr/local/bin/bundle "$@"
+    fix_file_permissions
+  fi
 fi
 
 if [ $command = "r" ]; then # rails
@@ -71,34 +83,54 @@ if [ $command = "r" ]; then # rails
   else
     executable=rails
   fi
-  docker_do run -it --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec $executable "$@"
+  if [ -e fig.yml ]; then
+    fig run --rm web bundle exec $executable $@
+  else
+    docker_do run -it --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec $executable "$@"
+  fi
   fix_file_permissions
 fi
 
 if [ $command = "s" ]; then # rails server
-  if [ -n "$rails_version" ] && [ ${rails_version%.*} = "2" ]; then
-    executable='ruby ./script/server'
+  if [ -e fig.yml ]; then
+    fig up
   else
-    executable='rails server'
+    if [ -n "$rails_version" ] && [ ${rails_version%.*} = "2" ]; then
+      executable='ruby ./script/server'
+    else
+      executable='rails server'
+    fi
+    running_server_id=`docker_do ps | grep 3000/tcp | awk '{print $1}'`
+    if [ -n "$running_server_id" ]; then
+      docker_do stop $running_server_id
+    fi
+    sudo rm -f $directory/tmp/pids/server.pid
+    docker_do run -it --rm $db_link -p 3000:3000 $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec $executable
   fi
-  running_server_id=`docker_do ps | grep 3000/tcp | awk '{print $1}'`
-  if [ -n "$running_server_id" ]; then
-    docker_do stop $running_server_id
-  fi
-  sudo rm -f $directory/tmp/pids/server.pid
-  docker_do run -it --rm $db_link -p 3000:3000 $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec $executable
 fi
 
 if [ $command = "k" ]; then # rake
-  docker_do run -i --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec rake "$@"
+  if [ -e fig.yml ]; then
+    fig run --rm web bundle exec rake $@
+  else
+    docker_do run -i --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec rake "$@"
+  fi
 fi
 
 if [ $command = "t" ]; then # test
-  docker_do run -it --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec guard
+  if [ -e fig.yml ]; then
+    fig run --rm web bundle exec guard
+  else
+    docker_do run -it --rm $db_link $code_volume $extra --entrypoint /usr/local/bin/bundle $app exec guard
+  fi
 fi
 
 if [ $command = "bash" ]; then # bash
-  docker_do run -it --rm $db_link $code_volume $extra $app /bin/bash --login
+  if [ -e fig.yml ]; then
+    fig run --rm web bash
+  else
+    docker_do run -it --rm $db_link $code_volume $extra $app /bin/bash --login
+  fi
 fi
 
 if [ $command = "dbfetch" ]; then # dbfetch
@@ -139,14 +171,22 @@ if [ $command = "dbload" ]; then # dbload
     echo "done" >> dbload.sh
     echo "$mysql_connection < db.sql" >> dbload.sh
     cp $db_dump_directory/db.sql .
-    sudo docker run -i --rm $db_link $code_volume $extra $app ./dbload.sh
+    if [ -e fig.yml ]; then
+      fig run --rm web sh ./dbload.sh
+    else
+      docker_do run -i --rm $db_link $code_volume $extra $app ./dbload.sh
+    fi
     rm db.sql
   else
     echo "/usr/bin/psql $app --username=$db_username --host=db -t -c 'drop schema public cascade; create schema public;'" >> dbload.sh
     echo "/usr/bin/pg_restore --username=$db_username --host=db --no-acl --no-owner --jobs=2 --dbname=$app db.dump" >> dbload.sh
     chmod +x dbload.sh
     cp $db_dump_directory/db.dump .
-    sudo docker run -i --rm $db_link $code_volume -e PGPASSWORD=$db_password $extra $app ./dbload.sh
+    if [ -e fig.yml ]; then
+      fig run --rm web sh ./dbload.sh
+    else
+      docker_do run -i --rm $db_link $code_volume -e PGPASSWORD=$db_password $extra $app ./dbload.sh
+    fi
     rm db.dump
   fi
   rm dbload.sh
